tdlib: Move all boxed types to the tdlib module
This way these types are no longer scattered around the codebase.
