+++ src/session/components/mod.rs
@@ -2,4 +2,4 @@ mod avatar;
 mod message_entry;
 
 pub(crate) use self::avatar::Avatar;
-pub(crate) use self::message_entry::{BoxedFormattedText, MessageEntry};
+pub(crate) use self::message_entry::MessageEntry;
